+++ src/deepuq/__init__.py
@@ -1,5 +1,5 @@
-from . import methods, models
+from . import data, methods, models
 from ._version import __version__
 from .types import UQResult
 
-__all__ = ["__version__", "methods", "models", "UQResult"]
+__all__ = ["__version__", "data", "methods", "models", "UQResult"]